refactor(hub): extract agent WS, install, enrollment into agentws.go
Pure code move — no behavior change. Same package, same globals
(db, agents/agentsMu, upgrader, pendingCmds, machineLatency, etc.).

hub/main.go: 2904 → 2226 lines, hub/agentws.go: 705 lines.

Moved (still package main):
  - ConnectedAgent type
  - handleAgentWS (the big WS handler)
  - Install flow: handleCreateToken, handleInstallScript,
    publicAndWebsocketBase, bootstrapCACertCandidates,
    loadBootstrapCACert, buildInstallCommand,
    handleDownloadAgent, handleDownloadCACert
  - Token lifecycle: validateAgentToken, consumeToken,
    consumeTokenAndStoreCredential, generateFirstRunToken
  - Durable agent secrets: generateAgentSecret, storeAgentCredential,
    machineHasCredential, validateAgentSecret, revokeAgentCredential

go test -count=1 ./...  green 3 runs in a row.
