+++ hub/main.go
@@ -3,12 +3,9 @@ package main
 import (
 	"bytes"
 	"context"
-	cryptoRand "crypto/rand"
-	"crypto/sha256"
 	"crypto/tls"
 	"crypto/x509"
 	"database/sql"
-	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -17,7 +14,6 @@ import (
 	"net/http"
 	"net/url"
 	"os"
-	"path/filepath"
 	"strconv"
 	"strings"
 	"sync"
@@ -113,12 +109,6 @@ type CommandResponse struct {
 	Error   string `json:"error"`
 }
 
-// ConnectedAgent tracks a live WebSocket connection from an agent.
-type ConnectedAgent struct {
-	MachineID string
-	Conn      *websocket.Conn
-	WriteMu   sync.Mutex
-}
 
 // TerminalSession tracks an active terminal relay session.
 var (
@@ -585,272 +575,13 @@ func handleSetTags(c echo.Context) error {
 
 // --- Install Script + Token ---
 
-func handleCreateToken(c echo.Context) error {
-	ip := getRealIP(c)
-	if !rateLimiter.Allow("token_create", ip, 3) {
-		log.Printf("rate limit exceeded: token_create from %s", ip)
-		return c.JSON(http.StatusTooManyRequests, map[string]string{"error": "rate limit exceeded"})
-	}
 
-	token := uuid.New().String()
-	h := sha256.Sum256([]byte(token))
-	tokenHash := hex.EncodeToString(h[:])
-	expiresAt := time.Now().Add(15 * time.Minute)
 
-	_, err := db.Exec(`INSERT INTO tokens (token_hash, expires_at) VALUES (?, ?)`, tokenHash, expiresAt)
-	if err != nil {
-		return c.JSON(http.StatusInternalServerError, map[string]string{"error": err.Error()})
-	}
 
-	httpBase, wsBase := publicAndWebsocketBase(c)
-	command, caURL, caSHA256 := buildInstallCommand(httpBase, wsBase, token)
 
-	resp := map[string]interface{}{
-		"token":      token,
-		"expires_at": expiresAt.Format(time.RFC3339),
-		"command":    command,
-	}
-	if caURL != "" {
-		resp["ca_url"] = caURL
-		resp["ca_sha256"] = caSHA256
-	}
-	return c.JSON(http.StatusOK, resp)
-}
 
-func handleInstallScript(c echo.Context) error {
-	script := `#!/bin/bash
-set -euo pipefail
 
-echo "=== BloxOS Agent Installer ==="
 
-# Detect architecture.
-ARCH=$(uname -m)
-case "$ARCH" in
-  x86_64)  ARCH="amd64" ;;
-  aarch64) ARCH="arm64" ;;
-  *) echo "Unsupported architecture: $ARCH"; exit 1 ;;
-esac
-
-HUB="${BLOXOS_HUB:?BLOXOS_HUB must be set}"
-TOKEN="${BLOXOS_TOKEN:?BLOXOS_TOKEN must be set}"
-HUB_HTTP=$(echo "$HUB" | sed 's|^ws://|http://|; s|^wss://|https://|')
-CA_URL="${BLOXOS_CA_URL:-}"
-CA_SHA256="${BLOXOS_CA_SHA256:-}"
-
-curl_fetch() {
-  local url="$1"
-  shift
-  if [[ "$url" == https://* ]]; then
-    curl --proto '=https' --tlsv1.2 -fsSL "$@" "$url"
-  else
-    curl -fsSL "$@" "$url"
-  fi
-}
-
-curl_fetch_bootstrap() {
-  local url="$1"
-  shift
-  if [[ "$url" == https://* ]]; then
-    curl --proto '=https' --tlsv1.2 -fsSLk "$@" "$url"
-  else
-    curl -fsSL "$@" "$url"
-  fi
-}
-
-echo "Hub: $HUB_HTTP"
-echo "Arch: $ARCH"
-
-CA_CURL_ARGS=()
-AGENT_CA_ENV=""
-
-if [[ -n "$CA_URL" ]]; then
-  if [[ -z "$CA_SHA256" ]]; then
-    echo "BLOXOS_CA_SHA256 must be set when BLOXOS_CA_URL is provided"
-    exit 1
-  fi
-
-  echo "Bootstrapping trusted CA..."
-  TMP_CA=$(mktemp)
-  trap 'rm -f "$TMP_CA"' EXIT
-  curl_fetch_bootstrap "$CA_URL" -o "$TMP_CA"
-  ACTUAL_CA_SHA=$(sha256sum "$TMP_CA" | awk '{print $1}')
-  if [[ "$ACTUAL_CA_SHA" != "$CA_SHA256" ]]; then
-    echo "CA fingerprint mismatch"
-    echo "Expected: $CA_SHA256"
-    echo "Actual:   $ACTUAL_CA_SHA"
-    exit 1
-  fi
-
-  sudo mkdir -p /etc/bloxos
-  sudo install -m 0644 "$TMP_CA" /etc/bloxos/ca.crt
-  CA_CURL_ARGS+=(--cacert /etc/bloxos/ca.crt)
-  AGENT_CA_ENV='Environment="BLOXOS_CA_CERT=/etc/bloxos/ca.crt"'
-fi
-
-# Download agent binary.
-echo "Downloading agent binary..."
-curl_fetch "${HUB_HTTP}/download/agent?arch=${ARCH}" "${CA_CURL_ARGS[@]}" -o /tmp/bloxos-agent
-chmod +x /tmp/bloxos-agent
-
-# Create system user (if not exists).
-if ! id -u bloxos &>/dev/null; then
-  sudo useradd -r -s /usr/sbin/nologin bloxos || true
-fi
-
-# Install binary.
-sudo mv /tmp/bloxos-agent /usr/local/bin/bloxos-agent
-
-# Create credential directory for agent secret (post-enrollment).
-sudo mkdir -p /etc/bloxos
-sudo chmod 700 /etc/bloxos
-
-# Create systemd service.
-# The agent uses the token for initial enrollment only.
-# After enrollment, it stores a durable secret in /etc/bloxos/agent-secret
-# and uses that for all future connections.
-sudo tee /etc/systemd/system/bloxos-agent.service > /dev/null << SVCEOF
-[Unit]
-Description=BloxOS Agent
-After=network.target
-
-[Service]
-Type=simple
-User=root
-Environment="BLOXOS_HUB=${HUB}"
-Environment="BLOXOS_TOKEN=${TOKEN}"
-${AGENT_CA_ENV}
-ExecStart=/usr/local/bin/bloxos-agent
-Restart=always
-RestartSec=5
-
-[Install]
-WantedBy=multi-user.target
-SVCEOF
-
-# Enable and start.
-sudo systemctl daemon-reload
-sudo systemctl enable bloxos-agent
-sudo systemctl restart bloxos-agent
-
-echo "=== BloxOS Agent installed and running ==="
-echo "Check status: systemctl status bloxos-agent"
-`
-	return c.String(http.StatusOK, script)
-}
-
-func publicAndWebsocketBase(c echo.Context) (string, string) {
-	publicURL := os.Getenv("PUBLIC_URL")
-	if publicURL != "" {
-		httpBase := publicURL
-		wsBase := strings.Replace(publicURL, "https://", "wss://", 1)
-		wsBase = strings.Replace(wsBase, "http://", "ws://", 1)
-		return httpBase, wsBase
-	}
-
-	host := c.Request().Host
-	proto := "ws"
-	httpProto := "http"
-	if c.Request().TLS != nil {
-		proto = "wss"
-		httpProto = "https"
-	}
-	return fmt.Sprintf("%s://%s", httpProto, host), fmt.Sprintf("%s://%s", proto, host)
-}
-
-func bootstrapCACertCandidates() []string {
-	candidates := []string{}
-	if env := os.Getenv("BLOXOS_CA_CERT"); env != "" {
-		candidates = append(candidates, env)
-	}
-	if home, err := os.UserHomeDir(); err == nil && home != "" {
-		candidates = append(candidates, filepath.Join(home, ".local", "share", "caddy", "pki", "authorities", "local", "root.crt"))
-	}
-	candidates = append(candidates,
-		"/var/lib/caddy/.local/share/caddy/pki/authorities/local/root.crt",
-		"/var/lib/caddy/pki/authorities/local/root.crt",
-		"/root/.local/share/caddy/pki/authorities/local/root.crt",
-	)
-	return candidates
-}
-
-func loadBootstrapCACert() ([]byte, string, error) {
-	for _, path := range bootstrapCACertCandidates() {
-		if path == "" {
-			continue
-		}
-		data, err := os.ReadFile(path)
-		if err == nil {
-			return data, path, nil
-		}
-		if os.IsNotExist(err) {
-			continue
-		}
-		return nil, "", fmt.Errorf("read CA cert %s: %w", path, err)
-	}
-	return nil, "", os.ErrNotExist
-}
-
-func buildInstallCommand(httpBase, wsBase, token string) (command string, caURL string, caSHA256 string) {
-	curlFlags := "-fsSL"
-	if strings.HasPrefix(httpBase, "https://") {
-		if caPEM, caPath, err := loadBootstrapCACert(); err == nil {
-			caURL = httpBase + "/download/ca.crt"
-			sum := sha256.Sum256(caPEM)
-			caSHA256 = hex.EncodeToString(sum[:])
-			curlFlags = "-fsSLk"
-			log.Printf("install bootstrap: using CA cert %s", caPath)
-		} else if !os.IsNotExist(err) {
-			log.Printf("WARNING: install bootstrap CA unavailable: %v", err)
-		}
-	}
-
-	command = fmt.Sprintf("export BLOXOS_HUB=%s BLOXOS_TOKEN=%s", wsBase, token)
-	if caURL != "" {
-		command += fmt.Sprintf(" BLOXOS_CA_URL=%s BLOXOS_CA_SHA256=%s", caURL, caSHA256)
-	}
-	command += fmt.Sprintf("; curl %s %s/install.sh | bash", curlFlags, httpBase)
-	return command, caURL, caSHA256
-}
-
-func handleDownloadAgent(c echo.Context) error {
-	// Configurable binary path (Finding #7).
-	// Resolution order: env var -> relative to working dir -> standard install path -> 404.
-	binaryPath := ""
-	candidates := []string{
-		os.Getenv("BLOXOS_AGENT_BINARY"),
-		"./agent/bloxos-agent",
-		"/usr/local/bin/bloxos-agent",
-	}
-	for _, p := range candidates {
-		if p == "" {
-			continue
-		}
-		if _, err := os.Stat(p); err == nil {
-			binaryPath = p
-			break
-		}
-	}
-	if binaryPath == "" {
-		return c.JSON(http.StatusNotFound, map[string]string{"error": "agent binary not found; set BLOXOS_AGENT_BINARY env var"})
-	}
-	// Log requested arch for future multi-arch support.
-	arch := c.QueryParam("arch")
-	if arch != "" {
-		log.Printf("agent download: arch=%s (serving default binary)", arch)
-	}
-	return c.File(binaryPath)
-}
-
-func handleDownloadCACert(c echo.Context) error {
-	caPEM, _, err := loadBootstrapCACert()
-	if err != nil {
-		if os.IsNotExist(err) {
-			return c.JSON(http.StatusNotFound, map[string]string{"error": "CA certificate not configured; set BLOXOS_CA_CERT"})
-		}
-		return c.JSON(http.StatusInternalServerError, map[string]string{"error": err.Error()})
-	}
-	return c.Blob(http.StatusOK, "application/x-pem-file", caPEM)
-}
 
 // --- Existing Handlers ---
 
@@ -900,237 +631,6 @@ func handleHealth(c echo.Context) error {
 	return c.JSON(http.StatusOK, map[string]string{"status": "ok"})
 }
 
-func handleAgentWS(c echo.Context) error {
-	token := c.QueryParam("token")
-	agentSecret := c.QueryParam("secret")
-
-	// Auth priority: secret > token. If neither, reject.
-	if token == "" && agentSecret == "" {
-		return c.JSON(http.StatusUnauthorized, map[string]string{"error": "token or secret required"})
-	}
-
-	// Mode B: Agent reconnecting with durable secret.
-	var secretMachineID string
-	if agentSecret != "" {
-		var err error
-		secretMachineID, err = validateAgentSecret(agentSecret)
-		if err != nil {
-			log.Printf("agent secret validation failed: %v", err)
-			// If secret was provided but invalid, and no token fallback, reject.
-			if token == "" {
-				return c.JSON(http.StatusUnauthorized, map[string]string{"error": "invalid agent secret"})
-			}
-			// Fall through to token-based auth.
-			agentSecret = ""
-		}
-	}
-
-	// Mode A: Token-based enrollment (only if not already authenticated via secret).
-	var tokenHash string
-	tokenValidated := false
-	if agentSecret == "" && token != "" {
-		var err error
-		tokenHash, err = validateAgentToken(token)
-		if err != nil {
-			tokenHash = ""
-			log.Printf("agent token validation deferred (may be reconnecting agent): %v", err)
-		} else {
-			tokenValidated = true
-		}
-	}
-
-	ws, wsErr := upgrader.Upgrade(c.Response(), c.Request(), nil)
-	if wsErr != nil {
-		return wsErr
-	}
-	defer ws.Close()
-
-	// If authenticated via secret, we already know the machine_id.
-	var machineID string
-	agent := &ConnectedAgent{Conn: ws}
-	if secretMachineID != "" {
-		machineID = secretMachineID
-		agent.MachineID = machineID
-		agentsMu.Lock()
-		agents[machineID] = agent
-		agentsMu.Unlock()
-		log.Printf("agent authenticated via secret: machine_id=%s", machineID)
-	}
-
-	for {
-		_, msg, err := ws.ReadMessage()
-		if err != nil {
-			log.Printf("agent disconnected: %v", err)
-			if machineID != "" {
-				markOffline(machineID)
-				agentsMu.Lock()
-				delete(agents, machineID)
-				agentsMu.Unlock()
-			}
-			return nil
-		}
-
-		var envelope struct {
-			Type      string `json:"type"`
-			MachineID string `json:"machine_id"`
-		}
-		if err := json.Unmarshal(msg, &envelope); err != nil {
-			log.Printf("invalid JSON from agent: %v", err)
-			continue
-		}
-
-		switch envelope.Type {
-		case "metrics", "":
-			var m AgentMetrics
-			if err := json.Unmarshal(msg, &m); err != nil {
-				log.Printf("invalid metrics JSON: %v", err)
-				continue
-			}
-			if machineID == "" {
-				machineID = m.MachineID
-				if secretMachineID != "" && machineID != secretMachineID {
-					log.Printf("rejecting agent: claimed machine_id %s does not match credential for %s", machineID, secretMachineID)
-					ws.WriteMessage(websocket.TextMessage, []byte(`{"type":"error","message":"machine_id does not match credential"}`))
-					return nil
-				}
-
-				// Check if this machine_id already exists (reconnecting agent).
-				var existingID string
-				knownMachine := db.QueryRow(`SELECT id FROM machines WHERE id = ?`, machineID).Scan(&existingID) == nil
-				hasCredential := machineHasCredential(machineID)
-
-				if knownMachine && secretMachineID == machineID {
-					log.Printf("known agent reconnecting with durable credential: %s (%s)", m.Hostname, machineID)
-				} else if tokenValidated {
-					// New enrollment with valid token - consume it and issue durable secret.
-					rawSecret, secretHash, err := generateAgentSecret()
-					if err != nil {
-						log.Printf("failed to generate agent secret: %v", err)
-						ws.WriteMessage(websocket.TextMessage, []byte(`{"type":"error","message":"enrollment failed"}`))
-						return nil
-					}
-					if err := consumeTokenAndStoreCredential(tokenHash, machineID, secretHash); err != nil {
-						log.Printf("failed to store agent credential: %v", err)
-						ws.WriteMessage(websocket.TextMessage, []byte(`{"type":"error","message":"enrollment failed"}`))
-						return nil
-					}
-					enrolledMsg, _ := json.Marshal(map[string]string{
-						"type":         "enrolled",
-						"agent_secret": rawSecret,
-					})
-					ws.WriteMessage(websocket.TextMessage, enrolledMsg)
-					log.Printf("new agent enrolled with durable secret: %s (%s)", m.Hostname, machineID)
-				} else if knownMachine || hasCredential {
-					log.Printf("rejecting known agent %s (%s): durable credential required", m.Hostname, machineID)
-					ws.WriteMessage(websocket.TextMessage, []byte(`{"type":"error","message":"durable credential required"}`))
-					return nil
-				} else {
-					log.Printf("rejecting unknown agent %s (%s): no valid token", m.Hostname, machineID)
-					ws.WriteMessage(websocket.TextMessage, []byte(`{"type":"error","message":"invalid or used token"}`))
-					return nil
-				}
-
-				agent.MachineID = machineID
-				agentsMu.Lock()
-				agents[machineID] = agent
-				agentsMu.Unlock()
-				log.Printf("agent registered: %s (%s)", m.Hostname, machineID)
-			}
-
-			// Calculate latency from sent_at.
-			if m.SentAt != "" {
-				sentTime, err := time.Parse(time.RFC3339Nano, m.SentAt)
-				if err == nil {
-					latencyMs := time.Since(sentTime).Milliseconds()
-					if latencyMs < 0 {
-						latencyMs = 0
-					}
-					machineLatencyMu.Lock()
-					machineLatency[m.MachineID] = latencyMs
-					machineLatencyMu.Unlock()
-				}
-			}
-
-			upsertMachine(m)
-			storeMetrics(m)
-
-			// Enrich the metrics broadcast with latency.
-			machineLatencyMu.RLock()
-			lat := machineLatency[m.MachineID]
-			machineLatencyMu.RUnlock()
-
-			// Re-marshal with latency included.
-			enriched := make(map[string]interface{})
-			json.Unmarshal(msg, &enriched)
-			enriched["latency_ms"] = lat
-			enrichedData, _ := json.Marshal(enriched)
-			broadcastSSE(enrichedData)
-
-		case "services":
-			var sm ServicesMessage
-			if err := json.Unmarshal(msg, &sm); err != nil {
-				log.Printf("invalid services JSON: %v", err)
-				continue
-			}
-			mid := sm.MachineID
-			if mid == "" {
-				mid = machineID
-			}
-			upsertServices(mid, sm.Services)
-			broadcastSSE(msg)
-
-		case "containers":
-			var cm ContainersMessage
-			if err := json.Unmarshal(msg, &cm); err != nil {
-				log.Printf("invalid containers JSON: %v", err)
-				continue
-			}
-			mid := cm.MachineID
-			if mid == "" {
-				mid = machineID
-			}
-			upsertContainers(mid, cm.Containers)
-			broadcastSSE(msg)
-
-		case "hardware_info":
-			// Static hardware snapshot — store the raw JSON as-is so adding
-			// new fields doesn't require another hub deploy.
-			var hw struct {
-				MachineID string `json:"machine_id"`
-			}
-			if err := json.Unmarshal(msg, &hw); err != nil {
-				log.Printf("invalid hardware_info JSON: %v", err)
-				continue
-			}
-			mid := hw.MachineID
-			if mid == "" {
-				mid = machineID
-			}
-			if _, err := db.Exec(`UPDATE machines SET hardware_info = ? WHERE id = ?`, string(msg), mid); err != nil {
-				log.Printf("store hardware_info: %v", err)
-			}
-
-		case "command_response":
-			var resp CommandResponse
-			if err := json.Unmarshal(msg, &resp); err != nil {
-				log.Printf("invalid command_response JSON: %v", err)
-				continue
-			}
-			pendingCmdsMu.Lock()
-			ch, ok := pendingCmds[resp.ID]
-			if ok {
-				delete(pendingCmds, resp.ID)
-			}
-			pendingCmdsMu.Unlock()
-			if ok {
-				ch <- resp
-			}
-
-		default:
-			log.Printf("unknown message type from agent: %s", envelope.Type)
-		}
-	}
-}
 
 func upsertServices(machineID string, services []ServiceInfo) {
 	tx, err := db.Begin()
@@ -1682,181 +1182,14 @@ func getMachinesJSON() ([]byte, error) {
 
 // --- Security Functions ---
 
-// validateAgentToken checks a token against the DB (Finding #1).
-// Returns the token hash so the caller can mark it as used after enrollment.
-func validateAgentToken(token string) (string, error) {
-	h := sha256.Sum256([]byte(token))
-	tokenHash := hex.EncodeToString(h[:])
 
-	var expiresAt string
-	var used bool
-	err := db.QueryRow(`SELECT expires_at, used FROM tokens WHERE token_hash = ?`, tokenHash).Scan(&expiresAt, &used)
-	if err == sql.ErrNoRows {
-		return "", fmt.Errorf("invalid token")
-	}
-	if err != nil {
-		return "", fmt.Errorf("database error: %w", err)
-	}
-	if used {
-		return "", fmt.Errorf("token already used")
-	}
-	expTime, err := time.Parse("2006-01-02 15:04:05", expiresAt)
-	if err != nil {
-		expTime, err = time.Parse(time.RFC3339, expiresAt)
-		if err != nil {
-			return "", fmt.Errorf("invalid expiry format")
-		}
-	}
-	if time.Now().After(expTime) {
-		return "", fmt.Errorf("token expired")
-	}
 
-	log.Printf("agent token validated successfully")
-	return tokenHash, nil
-}
 
-// consumeToken marks a token as used after successful enrollment.
-func consumeToken(tokenHash string) {
-	_, err := db.Exec(`UPDATE tokens SET used = TRUE WHERE token_hash = ?`, tokenHash)
-	if err != nil {
-		log.Printf("failed to mark token as used: %v", err)
-	} else {
-		log.Printf("install token consumed")
-	}
-}
 
-func consumeTokenAndStoreCredential(tokenHash, machineID, secretHash string) error {
-	tx, err := db.Begin()
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback()
 
-	res, err := tx.Exec(`UPDATE tokens SET used = TRUE WHERE token_hash = ? AND used = FALSE`, tokenHash)
-	if err != nil {
-		return err
-	}
-	rows, err := res.RowsAffected()
-	if err != nil {
-		return err
-	}
-	if rows != 1 {
-		return fmt.Errorf("install token already consumed")
-	}
 
-	if _, err := tx.Exec(`INSERT OR REPLACE INTO agent_credentials (machine_id, secret_hash, created_at, last_used_at)
-		VALUES (?, ?, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)`, machineID, secretHash); err != nil {
-		return err
-	}
-	return tx.Commit()
-}
 
-// generateAgentSecret creates a 32-byte random secret for agent enrollment.
-// Returns the hex-encoded raw secret and its SHA-256 hash.
-func generateAgentSecret() (raw string, hash string, err error) {
-	secretBytes := make([]byte, 32)
-	if _, err := cryptoRand.Read(secretBytes); err != nil {
-		return "", "", fmt.Errorf("generate random secret: %w", err)
-	}
-	raw = hex.EncodeToString(secretBytes)
-	h := sha256.Sum256([]byte(raw))
-	hash = hex.EncodeToString(h[:])
-	return raw, hash, nil
-}
 
-// storeAgentCredential saves (or replaces) the hashed secret for a machine.
-func storeAgentCredential(machineID, secretHash string) error {
-	_, err := db.Exec(`INSERT OR REPLACE INTO agent_credentials (machine_id, secret_hash, created_at, last_used_at)
-		VALUES (?, ?, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)`, machineID, secretHash)
-	return err
-}
-
-func machineHasCredential(machineID string) bool {
-	var count int
-	if err := db.QueryRow(`SELECT COUNT(*) FROM agent_credentials WHERE machine_id = ?`, machineID).Scan(&count); err != nil {
-		log.Printf("failed checking agent credential for %s: %v", machineID, err)
-		return false
-	}
-	return count > 0
-}
-
-// validateAgentSecret looks up a credential by the SHA-256 hash of the provided secret.
-// Returns the associated machine_id if found.
-func validateAgentSecret(secret string) (string, error) {
-	h := sha256.Sum256([]byte(secret))
-	secretHash := hex.EncodeToString(h[:])
-
-	var machineID string
-	err := db.QueryRow(`SELECT machine_id FROM agent_credentials WHERE secret_hash = ?`, secretHash).Scan(&machineID)
-	if err == sql.ErrNoRows {
-		return "", fmt.Errorf("invalid agent secret")
-	}
-	if err != nil {
-		return "", fmt.Errorf("database error: %w", err)
-	}
-
-	// Update last_used_at.
-	_, _ = db.Exec(`UPDATE agent_credentials SET last_used_at = CURRENT_TIMESTAMP WHERE secret_hash = ?`, secretHash)
-
-	return machineID, nil
-}
-
-// revokeAgentCredential removes the stored credential for a machine.
-func revokeAgentCredential(machineID string) error {
-	_, err := db.Exec(`DELETE FROM agent_credentials WHERE machine_id = ?`, machineID)
-	return err
-}
-
-// generateFirstRunToken creates a one-time token on first startup when no tokens and no machines exist.
-func generateFirstRunToken() {
-	var tokenCount int
-	if err := db.QueryRow(`SELECT COUNT(*) FROM tokens`).Scan(&tokenCount); err != nil {
-		log.Printf("first-run check: token count error: %v", err)
-		return
-	}
-	if tokenCount > 0 {
-		return
-	}
-
-	var machineCount int
-	if err := db.QueryRow(`SELECT COUNT(*) FROM machines`).Scan(&machineCount); err != nil {
-		log.Printf("first-run check: machine count error: %v", err)
-		return
-	}
-	if machineCount > 0 {
-		return
-	}
-
-	// First run: generate a real token with 1-hour expiry.
-	token := uuid.New().String()
-	h := sha256.Sum256([]byte(token))
-	tokenHash := hex.EncodeToString(h[:])
-	expiresAt := time.Now().Add(1 * time.Hour)
-
-	_, err := db.Exec(`INSERT INTO tokens (token_hash, expires_at) VALUES (?, ?)`, tokenHash, expiresAt)
-	if err != nil {
-		log.Printf("first-run: failed to insert token: %v", err)
-		return
-	}
-
-	// Write token to file instead of logging it (security hardening).
-	homeDir, err := os.UserHomeDir()
-	if err != nil {
-		log.Printf("first-run: cannot determine home dir: %v", err)
-		return
-	}
-	tokenDir := homeDir + "/.bloxos"
-	if err := os.MkdirAll(tokenDir, 0700); err != nil {
-		log.Printf("first-run: cannot create %s: %v", tokenDir, err)
-		return
-	}
-	tokenFile := tokenDir + "/first-run-token"
-	if err := os.WriteFile(tokenFile, []byte(token), 0600); err != nil {
-		log.Printf("first-run: cannot write token file: %v", err)
-		return
-	}
-	log.Printf("First-run token written to %s (expires in 1 hour)", tokenFile)
-}
 
 // tokenRedactingWriter redacts JWT tokens from log output (Finding #8).
 type tokenRedactingWriter struct {
